% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/land.R
\name{addurl}
\alias{addurl}
\title{Add URLs to a research project (land)}
\usage{
addurl(land_name, urls = NULL, path = NULL, db_name = "mwi.db")
}
\arguments{
\item{land_name}{A string specifying the name of the land.}

\item{urls}{A comma-separated string of URLs to add. Default is NULL.}

\item{path}{A string specifying the path to a file containing URLs. Default is NULL.}

\item{db_name}{A string specifying the name of the SQLite database file. Default is "mwi.db".}
}
\value{
Integer 1 if successful, 0 otherwise.
}
\description{
This function adds one or more URLs to a specified land in the SQLite database.
URLs can be provided directly as a comma-separated string, read from a file, or imported interactively.
}
\details{
The function performs the following steps:
1. Checks if the specified land exists in the Land table.
2. Retrieves the current date and time.
3. Adds the URLs from the provided list or file to the Expression table if they do not already exist.
4. Increments a counter for each new URL added and displays a message.
}
\examples{
\dontrun{
addurl("Climate Research", "http://example.com, http://example.org")
addurl("Climate Research", path = "urls.txt")
}
}
