% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crawl.R
\name{archive_available}
\alias{archive_available}
\title{Check URL availability in Archive.org}
\usage{
archive_available(url)
}
\arguments{
\item{url}{A character string representing the URL to be checked.}
}
\value{
A list containing the closest snapshot details if available, otherwise NULL.
}
\description{
This function checks if a given URL is available in Archive.org.
}
