% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{clean_string}
\alias{clean_string}
\title{Clean String}
\usage{
clean_string(str)
}
\arguments{
\item{str}{A character string to be cleaned.}
}
\value{
A cleaned character string.
}
\description{
This function removes control characters, newlines, and tabs from a string.
}
