% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crawl.R
\name{clean_url}
\alias{clean_url}
\title{Clean the extracted URL}
\usage{
clean_url(url)
}
\arguments{
\item{url}{A character string representing the URL to be cleaned.}
}
\value{
A cleaned URL as a character string.
}
\description{
This function removes fragment identifiers and keeps only valid URL characters.
}
