% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crawl.R
\name{crawlDetails}
\alias{crawlDetails}
\title{Crawl and extract details from a URL}
\usage{
crawlDetails(url)
}
\arguments{
\item{url}{A character string representing the URL to be crawled.}
}
\value{
A character string containing the extracted details.
}
\description{
This function crawls a URL and extracts detailed information such as meta descriptions,
Open Graph descriptions, Twitter descriptions, keywords, and header tags.
}
