% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crawl.R
\name{crawlDomain}
\alias{crawlDomain}
\title{Crawl and update the Domain table}
\usage{
crawlDomain(nburl = 100, db_name = "mwi.db")
}
\arguments{
\item{nburl}{An integer specifying the number of URLs to be crawled (default is 100).}

\item{db_name}{A character string representing the name of the database (default is "mwi.db").}
}
\description{
This function crawls domains and updates the Domain table with the fetched data.
}
