% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/land.R
\name{create_land}
\alias{create_land}
\title{Create a new research project (land)}
\usage{
create_land(name, desc, lang = "en", db_name = "mwi.db")
}
\arguments{
\item{name}{A string specifying the name of the land.}

\item{desc}{A string providing a description of the land.}

\item{lang}{A string specifying the language of the land. Default is "en".}

\item{db_name}{A string specifying the name of the SQLite database file. Default is "mwi.db".}
}
\value{
None. This function is called for its side effects.
}
\description{
This function creates a new entry in the Land table, representing a new research project.
If a land with the same name already exists, a message is displayed and no new entry is created.
}
\details{
The function first checks if a land with the same name already exists in the database.
If it does, a message is displayed. Otherwise, a new entry is inserted into the Land table
with the provided name, description, language, and the current timestamp.
}
\examples{
\dontrun{
create_land("Climate Research", "Research on climate change impacts", "en", "my_database.db")
}
}
