% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/land.R
\name{deleteland}
\alias{deleteland}
\title{Delete a research project (land)}
\usage{
deleteland(land_name, maxrel = NULL, db_name = "mwi.db")
}
\arguments{
\item{land_name}{A string specifying the name of the land to delete.}

\item{maxrel}{An integer specifying the maximum relevance for expressions to delete. Default is NULL.}

\item{db_name}{A string specifying the name of the SQLite database file. Default is "mwi.db".}
}
\value{
Integer 1 if successful, 0 otherwise.
}
\description{
This function deletes a specified land and its associated records from the SQLite database.
If a maximum relevance is specified, only the expressions with relevance less than or equal to this value are deleted.
}
\details{
The function performs the following steps:
1. Checks if the specified land exists in the Land table.
2. If a maximum relevance is specified, deletes expressions with relevance less than or equal to this value.
3. If no maximum relevance is specified, deletes the land and all its associated records.
}
\examples{
\dontrun{
deleteland("Climate Research")
deleteland("Climate Research", maxrel = 2)
}
}
