% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/land.R
\name{.get_land_id}
\alias{.get_land_id}
\title{Return the id of a land (helper)}
\usage{
.get_land_id(con, land_name)
}
\arguments{
\item{con}{A DBI connection.}

\item{land_name}{Character scalar.}
}
\value{
Integer id.
}
\description{
Internal utility returning the numeric id of a land, stops if it does not exist.
}
\keyword{internal}
