% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crawl.R
\docType{data}
\name{.trafilatura_mod}
\alias{.trafilatura_mod}
\title{Crawl a URL and extract content}
\format{
An object of class \code{python.builtin.module} (inherits from \code{python.builtin.object}) of length 1.
}
\usage{
.trafilatura_mod
}
\arguments{
\item{url}{A character string representing the URL to be crawled.}
}
\value{
A data frame containing the extracted content from the URL.
}
\description{
This function attempts to crawl a given URL using various methods, including
mercury-parser, trafilatura, and alternative methods to extract content.
It handles different content types, including PDFs.
}
\keyword{datasets}
