% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export_nodecsv}
\alias{export_nodecsv}
\title{Export Node CSV}
\usage{
export_nodecsv(con, land_name, minimum_relevance, filename)
}
\arguments{
\item{con}{A database connection object.}

\item{land_name}{A character string specifying the name of the land.}

\item{minimum_relevance}{A numeric value specifying the minimum relevance score for inclusion in the export.}

\item{filename}{A character string specifying the name of the output CSV file.}
}
\value{
An integer indicating the number of rows written to the CSV file.
}
\description{
This function exports domain data to a CSV file, including the number of expressions and average relevance score.
}
