% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{find_clusters}
\alias{find_clusters}
\title{Identifier des clusters dans un vecteur numérique avec modèles de mélange gaussiens}
\usage{
find_clusters(x, max_G = 9, auto_select_model = TRUE, modelNames = "V")
}
\arguments{
\item{x}{Un vecteur numérique représentant les données à segmenter en clusters.}

\item{max_G}{Un entier spécifiant le nombre maximum de clusters à tester lors de la modélisation.}

\item{auto_select_model}{Un booléen (TRUE/FALSE). Si TRUE, la fonction utilise le critère BIC pour laisser Mclust choisir automatiquement le meilleur modèle de structure de covariance. Si FALSE, le modèle est forcé à celui spécifié dans `modelNames`.}

\item{modelNames}{Une chaîne de caractères ou un vecteur de modèles à tester si `auto_select_model` est FALSE. Ces modèles définissent la structure de covariance utilisée dans le clustering.}
}
\value{
Retourne une liste contenant plusieurs éléments importants :
- `$fit_object` : L'objet complet retourné par Mclust, utile pour visualisations et diagnostics.
- `$best_model` : Le nom du modèle de covariance sélectionné comme meilleur.
- `$n_clusters` : Le nombre optimal de clusters détectés.
- `$classification` : Un vecteur indiquant l'appartenance de chaque observation à un cluster.
}
\description{
Cette fonction utilise des modèles de mélange gaussiens pour identifier des groupes latents (clusters) dans un vecteur numérique. Elle teste plusieurs modèles et nombres de clusters pour trouver la meilleure segmentation selon le critère BIC.
}
\details{
Il est souvent recommandé d'appliquer le clustering sur une variable transformée (par exemple, normalisée ou log-transformée) afin d'atténuer l'influence des valeurs aberrantes (outliers) qui pourraient fausser la segmentation.
}
