% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listurl.R
\name{get_serp_key}
\alias{get_serp_key}
\title{Retrieve SerpAPI key}
\usage{
get_serp_key()
}
\value{
Character scalar containing the API key.
}
\description{
Internal helper returning the key from the global variable or environment; if
missing, prompts the user.
}
\keyword{internal}
