% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{GPT_Recode}
\alias{GPT_Recode}
\title{GPT_Recode}
\usage{
GPT_Recode(
  prompt,
  cell,
  sysprompt = "You are a helpful assistant.",
  model = "gpt-3.5-turbo",
  temperature = 0.2
)
}
\arguments{
\item{prompt}{A character string. The prompt to provide to the model.}

\item{cell}{A character string. The cell value to be recoded.}

\item{sysprompt}{A character string. The system prompt to guide the model's behavior. Default is "You are a helpful assistant.".}

\item{model}{A character string. The model to use for the completion. Default is "gpt-3.5-turbo". Most famous are "gpt-4o"}

\item{temperature}{A numeric value. The temperature setting for the model's response. Default is 0.2.}
}
\value{
A character string. The recoded cell value based on the provided prompt.
}
\description{
A function to recode a cell value using the GPT-3.5-turbo model by providing a prompt.
}
\examples{
\dontrun{
  # Load the necessary library
  library(openai)

  # Set the API key for OpenAI
  Sys.setenv(OPENAI_API_KEY = "sk-proj-XXXXXXXXXXXX")

  # Use the GPT_Recode function
  result <- GPT_Recode(prompt = "Translate to French", cell = "Hello, how are you?")
  print(result)
}
}
