% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/land.R
\name{importFile}
\alias{importFile}
\title{Import URLs from a file}
\usage{
importFile()
}
\value{
Data frame containing the imported URLs.
}
\description{
This function allows the user to choose a file and import URLs from it.
The file can be in CSV or TXT format.
}
\details{
The function performs the following steps:
1. Allows the user to choose a file interactively.
2. Detects the file extension.
3. Reads the file based on the extension (CSV or TXT).
4. Returns the data frame containing the imported URLs.
}
\examples{
\dontrun{
urls <- importFile()
}
}
