% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crawl.R
\name{is_media_link}
\alias{is_media_link}
\title{Determine if a Link Points to a Media File}
\usage{
is_media_link(link)
}
\arguments{
\item{link}{A character string representing the URL or link to be evaluated.}
}
\value{
Returns a logical value (TRUE or FALSE) indicating whether the provided link points to a media file, based on its file extension.
}
\description{
This function evaluates a given link to determine if it refers to a media file, such as images, videos, or audio, by inspecting the file extension. It helps in filtering or categorizing links based on media content.
}
