% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_setup.R
\name{is_python3}
\alias{is_python3}
\title{Check if Python is Version 3}
\usage{
is_python3(python_path)
}
\arguments{
\item{python_path}{Path to Python executable.}
}
\value{
Logical indicating if it's Python 3.
}
\description{
Check if Python is Version 3
}
\keyword{internal}
