% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/land.R
\name{list_domain}
\alias{list_domain}
\title{List domains for a research project (land)}
\usage{
list_domain(land_name, db_name = "mwi.db")
}
\arguments{
\item{land_name}{A string specifying the name of the land.}

\item{db_name}{A string specifying the name of the SQLite database file. Default is "mwi.db".}
}
\value{
A vector of domain names.
}
\description{
This function lists all unique domains associated with a specified land in the SQLite database.
}
\details{
The function performs the following steps:
1. Establishes a connection to the SQLite database.
2. Constructs and executes an SQL query to retrieve distinct domain names associated with the specified land.
3. Returns a vector of domain names.
}
\examples{
\dontrun{
list_domain("Climate Research")
}
}
