% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crawl.R
\name{mkdictionary}
\alias{mkdictionary}
\title{Create a dictionary from text}
\usage{
mkdictionary(text, language = "fr")
}
\arguments{
\item{text}{A character string representing the input text.}

\item{language}{A character string specifying the language for stemming (default is "fr").}
}
\value{
A data frame containing the stemmed phrases as the dictionary.
}
\description{
This function creates a dictionary by tokenizing and stemming phrases from the input text.
}
