% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crawl.R
\name{PDFtoText}
\alias{PDFtoText}
\title{Extract Text Content from a PDF URL}
\usage{
PDFtoText(url)
}
\arguments{
\item{url}{A character string representing the URL where the PDF file is located.}
}
\value{
Returns a character string containing the extracted text content from the PDF file.
}
\description{
This function downloads a PDF document from the specified URL and attempts to extract its textual content. It primarily uses the `pdftools` package for text extraction and falls back to Optical Character Recognition (OCR) if the text extraction fails or is incomplete.
}
