% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_setup.R
\name{print.mwi_python_status}
\alias{print.mwi_python_status}
\title{Print Python Status}
\usage{
\method{print}{mwi_python_status}(x, ...)
}
\arguments{
\item{x}{A mwi_python_status object.}

\item{...}{Additional arguments (ignored).}
}
\description{
Print Python Status
}
