% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listurl.R
\name{related_query}
\alias{related_query}
\title{Retrieve Related Search Queries from Google using SerpAPI}
\usage{
related_query(query, lang = "fr", country = "France")
}
\arguments{
\item{query}{A character string representing the search query.}

\item{lang}{A character string representing the language for the search (default is "fr").}

\item{country}{A character string representing the country for the search (default is "France").}
}
\value{
A data frame containing the related searches, with columns:
\describe{
  \item{block_position}{Integer indicating the position of the block in the search results.}
  \item{query}{Character string of the related query.}
  \item{link}{Character string of the link to the related query.}
  \item{serpapi_link}{Character string of the SerpAPI link for the related query.}
}
If no related searches are found, the function returns NULL.
}
\description{
This function retrieves related search queries from Google using the SerpAPI service.
It constructs an API request based on the provided query, language, and country parameters,
then processes the response to extract and return the related searches.
}
\examples{
\dontrun{
related_searches <- related_query("data science", lang = "en", country = "United States")
}

}
