% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{slugify}
\alias{slugify}
\title{Slugify a String}
\usage{
slugify(string)
}
\arguments{
\item{string}{A character string to be slugified.}
}
\value{
A slugified character string.
}
\description{
This function creates a URL-friendly slug from a given string.
}
