% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/land.R
\name{stem_word}
\alias{stem_word}
\title{Stem a word}
\usage{
stem_word(word, language = "en")
}
\arguments{
\item{word}{A string specifying the word to stem.}

\item{language}{A string specifying the language for stemming. Default is "en".}
}
\value{
The stemmed word.
}
\description{
This function stems a word using the specified language.
}
\details{
The function converts the word to lowercase and then applies a stemming algorithm based on the specified language.
}
\examples{
\dontrun{
stem_word("running", "en")
}
}
