% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{to_metadata}
\alias{to_metadata}
\title{Convert to Metadata}
\usage{
to_metadata(lignes)
}
\arguments{
\item{lignes}{A data frame row containing the fields for metadata.}
}
\value{
A character string containing the metadata.
}
\description{
This function creates a metadata string from the given data frame row.
}
