% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listurl.R
\name{urlist_Bing}
\alias{urlist_Bing}
\title{Retrieve URLs with SerpAPI Bing Search}
\usage{
urlist_Bing(query, filename = NULL, sleep_seconds = 1, lang = "fr")
}
\arguments{
\item{query}{A character string representing the search query.}

\item{filename}{A character string representing the name of the file where the results will be saved.
If NULL, a filename will be generated based on the query (default is NULL).}

\item{sleep_seconds}{A numeric value representing the sleep time between API requests to avoid rate-limiting (default is 1).}

\item{lang}{A character string representing the language for the search (default is "fr").}
}
\value{
This function does not return a value. It writes the search results to a text file.
}
\description{
This function retrieves URLs from Bing search results using the SerpAPI service.
It constructs API requests based on the provided query and language parameters,
then processes the response to extract and save the URLs in a text file.
}
\examples{
\dontrun{
urlist_Bing("data science", filename = "data_science_results.txt", sleep_seconds = 2, lang = "en")
}

}
