% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listurl.R
\name{urlist_Duck}
\alias{urlist_Duck}
\title{Retrieve URLs from DuckDuckGo Search Using SerpAPI}
\usage{
urlist_Duck(query, filename = NULL, sleep_seconds = 1, lang = "fr")
}
\arguments{
\item{query}{A character string representing the search query to be sent to DuckDuckGo.}

\item{filename}{A character string specifying the name of the file where the search results will be saved. If NULL, a filename is automatically generated based on the query.}

\item{sleep_seconds}{A numeric value specifying the delay in seconds between API requests to prevent rate limiting. Default is 1 second.}

\item{lang}{A character string specifying the language code for the search. Default is "fr" (French).}
}
\value{
This function does not return a value. It writes the retrieved URLs from DuckDuckGo search results to a text file.
}
\description{
This function uses the SerpAPI service to perform a DuckDuckGo search with the specified query and language. It processes the API response to extract URLs from the search results and saves them into a text file for further use.
}
\examples{
\dontrun{
# Retrieve URLs for the query "data science" and save to a file
urlist_Duck("data science", filename = "data_science_results.txt", sleep_seconds = 2, lang = "en")
}
}
