% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crawl.R
\name{word_tokenizer}
\alias{word_tokenizer}
\title{Tokenize words}
\usage{
word_tokenizer(text)
}
\arguments{
\item{text}{A character string representing the text to be tokenized.}
}
\value{
A character vector of tokenized words.
}
\description{
This function splits a text into words.
}
