% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/land.R
\name{addterm}
\alias{addterm}
\title{Add terms to a research project (land)}
\usage{
addterm(land_name, terms)
}
\arguments{
\item{land_name}{A string specifying the name of the land.}

\item{terms}{A comma-separated string of terms to add.}
}
\value{
Integer 1 if successful, 0 otherwise.
}
\description{
This function adds one or more terms to a specified land in the SQLite database.
If the terms do not already exist in the Word table, they are added along with their stemmed lemmas.
If the terms already exist, they are simply associated with the specified land in the LandDictionary table.
}
\details{
The function performs the following steps:
1. Checks if the specified land exists in the Land table.
2. Splits the terms string into individual terms.
3. Converts each term to lowercase and checks if it already exists in the Word table.
4. If a term does not exist, it is stemmed and added to the Word table along with its lemma.
5. Each term is then associated with the specified land in the LandDictionary table.
}
\examples{
\dontrun{
addterm("Climate Research", "global warming", "climate", "change")
}
}
