% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{annotatedData}
\alias{annotatedData}
\title{Update Database Table with Externally Modified Data}
\usage{
annotatedData(dataplus, table, champ, by, labase = "mwi.db")
}
\arguments{
\item{dataplus}{A data frame containing the modified data to be inserted into the database.}

\item{table}{A character string specifying the name of the table to be updated in the database.}

\item{champ}{A character string specifying the name of the field (column) to be updated.}

\item{by}{A character string specifying the name of the key field used to match rows for updating.}

\item{labase}{A character string specifying the name of the SQLite database file. Default is "mwi.db".}
}
\value{
This function does not return a value. It updates the specified table in the database
and prints a success message upon completion.
}
\description{
This function allows reinsertion of externally modified data into the project's database.
It's particularly useful in collaborative or open science workflows where data might be
modified outside the main project environment, such as by collaborators or in open science initiatives.
}
\details{
The function performs the following steps:
1. Establishes a connection to the specified SQLite database.
2. Verifies the existence of the specified table and fields.
3. Prepares an SQL UPDATE statement.
4. Executes the update in a transaction for improved performance and data integrity.
5. Commits the changes if successful, or rolls back if an error occurs.

This function is designed to facilitate workflows where data might be exported, modified
externally (e.g., by collaborators or in spreadsheet software), and then reintegrated
into the main project database. It ensures that externally processed data can be
seamlessly incorporated back into the project's central data store.
}
\note{
- Ensure that the structure of `dataplus` matches the database table, particularly
  the columns specified by `champ` and `by`.
- The function uses transactions, so either all updates are applied, or none are
  (in case of an error), maintaining database consistency.
- It's recommended to backup your database before performing large-scale updates.
}
\examples{
\dontrun{
# Assuming 'modified_data' is a data frame with updated information
AnnotatedData(dataplus = modified_data,
              table = "expression",
              champ = "description",
              by = "id",
              labase = "mwi.db")
}

}
