% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{connect_db}
\alias{connect_db}
\title{Establish a Database Connection}
\usage{
connect_db(db_name = "mwi.db")
}
\arguments{
\item{db_name}{A character string specifying the name of the database file. Default is "mwi.db".}
}
\value{
A database connection object.
}
\description{
This function establishes a connection to the specified SQLite database.
}
