% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crawl.R
\name{crawl}
\alias{crawl}
\title{Crawl a URL and extract content}
\usage{
crawl(url)
}
\arguments{
\item{url}{A character string representing the URL to be crawled.}
}
\value{
A data frame containing the extracted content from the URL.
}
\description{
This function attempts to crawl a given URL using various methods, including
mercury-parser, trafilatura, and alternative methods to extract content.
It handles different content types, including PDFs.
}
