% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export_fullpagecsv}
\alias{export_fullpagecsv}
\title{Exporter les données complètes d'une page au format CSV}
\usage{
export_fullpagecsv(con, land_name, minimum_relevance, filename)
}
\arguments{
\item{con}{Objet de connexion à la base de données permettant d'accéder aux données du territoire.}
\item{land_name}{Chaîne de caractères indiquant le nom du territoire (land) dont on souhaite exporter les données.}
\item{minimum_relevance}{Valeur numérique définissant le seuil minimal de pertinence pour inclure une page dans l'export. Seules les pages avec une pertinence égale ou supérieure seront exportées.}
\item{filename}{Nom du fichier CSV de sortie dans lequel les données seront enregistrées.}
}
\value{
Un entier indiquant le nombre de lignes (pages) écrites dans le fichier CSV.
}
\description{
Cette fonction permet d'exporter les données détaillées des pages d'un territoire donné au format CSV. 
Les données sont extraites de la base via la connexion \code{con}, filtrées selon le seuil de pertinence \code{minimum_relevance}, puis sauvegardées dans un fichier CSV nommé \code{filename}.
Cette exportation facilite l'analyse externe, le partage ou l'archivage des données complètes des pages.
}
