% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export_land}
\alias{export_land}
\title{Export Land Data}
\usage{
export_land(
  land_name,
  export_type,
  minimum_relevance = 1,
  labase = "mwi.db",
  ext = "md"
)
}
\arguments{
\item{land_name}{A character string specifying the name of the land.}

\item{export_type}{A character string specifying the type of export ("pagecsv", "fullpagecsv", "nodecsv", "mediacsv", "pagegexf", "nodegexf", or "corpus").}

\item{minimum_relevance}{A numeric value specifying the minimum relevance score for inclusion in the export. Default is 1.}

\item{labase}{A character string specifying the name of the database file. Default is "mwi.db".}

\item{ext}{A character string specifying the file extension for corpus export ("md", "txt", or "pdf"). Default is "md". PDF requires rmarkdown and pandoc.}
}
\description{
This function manages the exportation of land data based on the specified export type.
}
