% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export_pagegexf}
\alias{export_pagegexf}
\title{Export Page GEXF}
\usage{
export_pagegexf(con, land_name, minimum_relevance, filename)
}
\arguments{
\item{con}{A database connection object.}

\item{land_name}{A character string specifying the name of the land.}

\item{minimum_relevance}{A numeric value specifying the minimum relevance score for inclusion in the export.}

\item{filename}{A character string specifying the name of the output GEXF file.}
}
\description{
This function exports expression data from the specified land to a GEXF file for network analysis.
}
