% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crawl.R
\name{expression_relevance}
\alias{expression_relevance}
\title{Calculate expression relevance}
\usage{
expression_relevance(dictionary, expression, language = "fr")
}
\arguments{
\item{dictionary}{A character vector of terms to be used for relevance calculation.}

\item{expression}{A data frame containing the title and text of the expression.}

\item{language}{A character string specifying the language for stemming (default is "fr").}
}
\value{
A numeric value representing the relevance score.
}
\description{
This function calculates the relevance of an expression based on its title and text.
}
