% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/land.R
\name{listlands}
\alias{listlands}
\title{List research projects (lands)}
\usage{
listlands(land_name = NULL, db_name = "mwi.db")
}
\arguments{
\item{land_name}{A string specifying the name of the land to list. If NULL, all lands are listed. Default is NULL.}

\item{db_name}{A string specifying the name of the SQLite database file. Default is "mwi.db".}
}
\value{
None. This function is called for its side effects.
}
\description{
This function lists all lands or a specific land from the SQLite database.
It retrieves and displays information such as the name, creation date, description,
terms in the dictionary, total number of expressions, number of expressions remaining to be fetched,
and HTTP status codes of the expressions.
}
\details{
The function retrieves and displays the following information for each land:
- Name of the land
- Creation date
- Description
- Terms in the dictionary
- Total number of expressions
- Number of expressions remaining to be fetched
- HTTP status codes of the expressions
}
\examples{
\dontrun{
listlands()
listlands("Climate Research", "my_database.db")
}
}
