% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crawl.R
\name{PDFtoText}
\alias{PDFtoText}
\title{Extract text from a PDF URL}
\usage{
PDFtoText(url)
}
\arguments{
\item{url}{A character string representing the URL of the PDF.}
}
\value{
A character string containing the extracted text from the PDF.
}
\description{
This function downloads a PDF from a given URL and attempts to extract
the text using `pdftools` and OCR if necessary.
}
