% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listurl.R
\name{related_query}
\alias{related_query}
\title{Retrieve Related Search Queries from Google Using SerpAPI}
\usage{
related_query(query, lang = "fr", country = "France")
}
\arguments{
\item{query}{A character string representing the main search query for which related queries are sought.}

\item{lang}{A character string specifying the language code for the search. Default is "fr" (French).}

\item{country}{A character string specifying the country name for the search context. Default is "France".}
}
\value{
Returns a data frame containing related search queries with the following columns:
\describe{
  \item{block_position}{An integer indicating the position of the related query block in the search results.}
  \item{query}{A character string representing the related search query.}
  \item{link}{A character string containing the URL link to the related query.}
  \item{serpapi_link}{A character string containing the SerpAPI URL for the related query.}
}
If no related searches are found, the function returns NULL.
}
\description{
This function uses the SerpAPI service to retrieve related search queries from Google based on a given main query. It constructs an API request using the specified query, language, and country parameters, then processes the API response to extract and return related search queries in a structured data frame.
}
\examples{
\dontrun{
# Retrieve related search queries for "data science" in English and United States context
related_searches <- related_query("data science", lang = "en", country = "United States")
}
}
