% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{slugify}
\alias{slugify}
\title{Convert a String into a URL-Friendly Slug}
\usage{
slugify(string)
}
\arguments{
\item{string}{A character string that will be transformed into a slug.}
}
\value{
Returns a character string that is a URL-friendly slug version of the input string. This typically involves converting to lowercase, replacing spaces with hyphens, and removing special characters.
}
\description{
This function transforms an input string into a slug suitable for use in URLs. It ensures the resulting string is lowercase, contains only URL-safe characters, and replaces spaces and other delimiters with hyphens to improve readability and SEO.
}
