% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{to_metadata}
\alias{to_metadata}
\title{Convert a Data Frame Row to a Metadata String}
\usage{
to_metadata(lignes)
}
\arguments{
\item{lignes}{A single row from a data frame containing fields to be converted into metadata.}
}
\value{
Returns a character string representing the metadata constructed from the input data frame row.
}
\description{
This function takes a row from a data frame and constructs a metadata string by extracting and formatting the relevant fields. The resulting string can be used for annotation, documentation, or further processing in data workflows.
}
