% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{transform_variable}
\alias{transform_variable}
\title{Transform a Numeric Vector for Variance Stabilization or Normalization}
\usage{
transform_variable(x, method = "yeojohnson", ...)
}
\arguments{
\item{x}{A numeric vector to be transformed.}

\item{method}{A character string specifying the transformation method to apply. Options include "none", "log1p", "log", "boxcox", and "yeojohnson".}

\item{...}{Additional arguments passed to the `bestNormalize` package functions (e.g., `standardize = FALSE`).}
}
\value{
Returns a list containing:
- `$val`: The transformed numeric vector.
- `$param`: An object containing the parameters of the transformation fit, useful for applying the same transformation to new data via `predict`.
}
\description{
This function applies a specified transformation to a numeric vector to stabilize variance or normalize its distribution. It supports multiple transformation methods and allows additional customization through extra arguments.
}
