% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listurl.R
\name{urlist_Google}
\alias{urlist_Google}
\title{Query to URLs with SerpAPI Google Search}
\usage{
urlist_Google(query, datestart, dateend, timestep = "week")
}
\arguments{
\item{query}{A character string representing the search query.}

\item{datestart}{A character string representing the start date for the search in "yyyy-mm-dd" format.}

\item{dateend}{A character string representing the end date for the search in "yyyy-mm-dd" format.}

\item{timestep}{A character string representing the time step for the date sequence (default is "week").}

\item{sleep_seconds}{A numeric value representing the sleep time between API requests to avoid rate-limiting (default is 1).}

\item{lang}{A character string representing the language for the search (default is "fr").}
}
\value{
This function does not return a value. It writes the search results to a text file.
}
\description{
This function retrieves URLs from Google search results using the SerpAPI service.
It constructs API requests based on the provided query, date range, and language parameters,
then processes the response to extract and save the URLs in a text file.
}
\examples{
\dontrun{
urlist_Google("data science", "2023-01-01", "2023-03-01", timestep = "month", sleep_seconds = 2, lang = "en")
}

}
